SPARKC-606 honor ContinuousPaging properties
Reintroduce ContinuousPaging throttle configuration properties.
Before this change the properties were silently ignored, this was a side
effect introduced by the Huge Java Driver Refactoring.

This PR also deprecates Scanner.getSession as it breaks encapsulation.
Ideally Scanner class should expose only scan method but this is not
possible as the callers need a way to prepare a scan statement.
New prepare method was introduced to get around this requirement.